refactoring code to organise it better show / hide widgets on welcome screen Implement setup panel using uic
