+++ .gitignore
@@ -75,3 +75,4 @@ app.py
 
 .shell.nix.swp
 geest.tar.gz
+gitlog.txt
