+++ geest/core/__init__.py
@@ -6,5 +6,7 @@ from .constants import APPLICATION_NAME
 from .settings import setting, set_setting
 
 from .default_settings import default_settings
+from .json_tree_item import JsonTreeItem
+from .workflow_queue_manager import WorkflowQueueManager
 
 # from .json_validator import JSONValidator
