+++ geest/core/workflows/dont_use_workflow.py
@@ -1,7 +1,7 @@
 import time
 from qgis.core import QgsMessageLog, Qgis, QgsFeedback
 from .workflow_base import WorkflowBase
-from geest.gui.treeview import JsonTreeItem
+from geest.gui.views import JsonTreeItem
 
 
 class DontUseWorkflow(WorkflowBase):
