+++ geest/core/workflows/multi_buffer_distances_workflow.py
@@ -17,7 +17,7 @@ from qgis.core import (
 from qgis.PyQt.QtCore import QVariant
 import processing  # QGIS processing toolbox
 from .workflow_base import WorkflowBase
-from geest.gui.treeview import JsonTreeItem
+from geest.gui.views import JsonTreeItem
 from geest.core.multibuffer_point import MultiBufferCreator
