+++ geest/gui/dialogs/__init__.py
@@ -1,0 +1,2 @@
+from .factor_aggregation_dialog import FactorAggregationDialog
+from .indicator_detail_dialog import IndicatorDetailDialog
