+++ geest/gui/dialogs/factor_aggregation_dialog.py
