+++ geest/gui/panels/__init__.py
@@ -1,0 +1,2 @@
+from .setup_panel import SetupPanel
+from .tree_panel import TreePanel
